Update `indices.flush`-fn docstrings for ES 2.3
I removed `:refresh` as an accepted option because Elasticsearch itself
[removed that flag from the Flush API][1].

N.B. the native implementation expects “kebab-case” (`:wait-if-ongoing`)
instead of “snake_case” (`:wait_if_ongoing`), and the docstring
descriptions reflect that. I used snake_case in the REST implementation
for consistency with the other REST client code. The implementations
can (and probably should) be aligned in a separate commit.

[1]: https://github.com/elastic/elasticsearch/commit/7cc48c8e8723d3b31fbcb371070bc2a8d87b1f7e
     "Flush API: remove refresh flag"