TODO UPDATE EARLIER ONE TO MATCH?? Fix `c.e.r.index/clear-cache` dropped-option bug
The values for `:filter`, `:field_data`, and `:bloom` _should_ have
been sent to Elasticsearch as “request parameters” (i.e., they should
have been in the URL, as query parameters) but they were being sent
in the _body_ of the request. I believe Elasticsearch disregarded
the values provided in the body and thus —in effect— missed
the Elastisch user’s intent.

Sources:
- <https://github.com/elastic/elasticsearch/blob/v0.90.13/rest-api-spec/api/indices.clear_cache.json>
  In particular, note how `filter`, `field_data` and `bloom` are all
  under `url.params`, and `body` is `null`.
- <https://github.com/elastic/elasticsearch/blob/v0.19.8/src/main/java/org/elasticsearch/rest/action/admin/indices/cache/clear/RestClearIndicesCacheAction.java#L62-L65>
  (the state of the relevant server-side code at the time of
  [`clear-cache`’s inception](https://github.com/clojurewerkz/elastisch/commit/b2a9c6bdb7d567e9a07191b59a023fc88a0de7f7)