Fix dropped-option bug in `c.e.r.index/flush`
The value for `:refresh` _should_ have been sent to Elasticsearch
as a “request parameter” —i.e., it should have been a part of the URL—
but it was being sent in the _body_ of the request. As a result,
Elasticsearch ignored the `refresh` data provided in the body and
effectively missed the Elastisch user’s intent.

Sources:
- <http://web.archive.org/web/20120730030655/http://www.elasticsearch.org/guide/reference/api/admin-indices-flush.html>
  The state of the ElasticSearch documentation at the time of
  [`flush`’s initial implementation](https://github.com/clojurewerkz/elastisch/commit/bb90f25a8549c707c6c63e40caf205e651cf1311#diff-8865449ff4f19fd95e480a636e676a05R166).
- <https://github.com/elastic/elasticsearch/blob/v0.90.13/rest-api-spec/api/indices.flush.json#L29-L32>